Can now disable the print() statements in AvailableWords()
Did some organization of the import statements and functions in
find_best_starting_words.py. Probable not any faster, but
calculations were changed to be generators to reduce memory
overhead.

Multiprocessing with Pool is not available since Available is class
and is not pickleable. This really needs some multiprocessing to
speed things up.