major bug fixes to calculate_best_words
Found bug where I was allowing impossible outcomes to be calculated.
This resulted multiple outcomes where the same words were possible.
Since when you type a word into wordle you don't get two outcomes
this is unphysical. Thus I added in a fuction to strip impossible
outcomes from the calculation. In order to help find what outcomes
where impossible I need to implament a checksum where you check that
for each guess word the total number of words in all outcomes equals
the total number of words that can be guessed. To allow checksum to
show what outcomes failed for what words I now have a option
to return and save the entire lists of words not just the number of
words in each outcome. tried to play several games with helper and
did not see the checksum trigger so I think the alogorith is working
now. Still need major clean up and orgainzation to all of additions.