added in decision memory for Jordan bot
Jordan is not the fastest. Thus is it important that Jordan doesn't
waste time thinking about problems he has already solved. For
instance raise will return 00000 168 times. There is no point in
in recalculating the best word to use everytime. In order to for
that to work I had to stop reinitializing the hint every word in
play. Also I had to pass the guessed_words and guessed_results to
the hint class. I think that this didn't break any other features.
It looks as though this was expected one day when hint.py was
written