new modes fastest and all
I added in new modes for the metric of calculate_best_words
the newest metric is what word will allow you to have the best
chance of getting the answer correct the next guess after choosing
that word. This is becuase if all you want is to post great wordle
squares on your insta then all you care about is getting it in two
or three guesses anything else you just rage quit.
Also new mode is all where it still returns the smalles average
list length but it displays all three of the current metrics.