has error in mode fastest probability math
metric for fastest was not correctly calculated for what I said
it should be doing. Also I changed hint.py to force the bot to
always pick possible answers after the first guess to match the
personality of trying to get it right in least number of guesses
at expense of have some really bad games.