added in new metrics for calculate_best_words
I added in a variance metric which caculates the variance in list
length. The idea is that best case would be splitting
all of the remaining words up evenly over all outcomes in which
case the list length would have a small variance.
Also added a guess cost metric. This uses data from bot runs on the
top 10 words with mode = ['split','var'] then looking at the
average cost in guesses for all the different sizes of list length
obtained during those runs. This data was then fit with a power law
to generate a guess cost for any size list.
Also now you can just suppy a list of modes you want to use and
all of those metrics will be displayed with the first being the
dominant metric and following metrics used to break ties.