Changed the game play.
Hints are no longer the default option, and to get hint now you
need to specify --hint or --auto-play options from the commandline.

Hints are no longer automatically played. Bot hints are printed to
the screen by Default. use --auto-play to auto-play the bot hints.

If hints are allowed (--hint is used) there is now another secret
tool, use the word 'narrow' as a guess, and it will print the string
from the AvailableWords() method used in the current game.

The bots Jada and Natalie needed upgrades make better selections
in and out of hard mode. The Natalie bots bias for letters starting
with 'A' was also fixed.