fixed issue with hard mode allowed guesses
There was a slight difference between how narrow and thus
play.py allowed second guesses for hard mode compared to
actual wordle. In actual wordle on hard mode if you get a yellow
tile you are allowed to have a guess with that same letter in that
same yellow tile as position even though this is not a possible solution.
Since narrow used the same engine to narrow allowed guesses as
it uses to narrow the possible remaining answers this was a problem.
Likewise if you guess a repeated letter and the tiles confirm
what the maximum allowed instances of the that repeated letter
are you are still allowed to guess that letter more times that the max
possible number of instances on hard mode. This really doesn't
matter most of the time because these are in general bad guesses
that minimize the amount of new information. but having a bit more
freedom to guess is slightly better for the Jordan bot in the same
way that normal mode can solve in less guesses than hard mode.
For the Jordan bot it reduced the total number of guesses in
hard mode by 1 for salet start word (so a very small difference).